Version 1 Complete
Corrected settings update information. Determined where the problem is coming from and will make note of it for current users and testing.
Added categories feature in "write" section. Posts will now be able to have a category assigned to it upon completion. Default category is set to "Code". Change to life in next update. Styled the categories button and moved the publish button to flow within the flex display to prep for mobile edition.
Adjusted margins for title and write section for mobile edition.
Note: When updating a user, all input fields must be modified or else a crash will occur. Must determine why the crash is occurring.